// src/shared/api/models/getApiSoldBundleOffersIdHistory200Item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiSoldBundleOffersIdHistory200ItemAppliedService } from './getApiSoldBundleOffersIdHistory200ItemAppliedService';
import type { GetApiSoldBundleOffersIdHistory200ItemEmployee } from './getApiSoldBundleOffersIdHistory200ItemEmployee';

export type GetApiSoldBundleOffersIdHistory200Item = {
    id?: string;
    bundleOfferId?: string;
    /** @nullable */
    salesId?: string | null;
    /** @nullable */
    invoiceNumber?: string | null;
    outletId?: number;
    punchesUsed?: number;
    /** Services on which bundle offer can be applied */
    appliedService?: GetApiSoldBundleOffersIdHistory200ItemAppliedService;
    punchesRemaining?: number;
    /** @nullable */
    usedBy?: number | null;
    /** @nullable */
    employee?: GetApiSoldBundleOffersIdHistory200ItemEmployee;
    createdAt?: string;
};